moved time + currency spinners to spinsliders.py
